// src/enums/asset-paths.enum.ts
export enum AssetPaths {
    FEELING_PROUD_SVG = "assets/illustrations/FeelingProud.svg",
    DIGITAL_SOLUTIONS_SVG = "assets/illustrations/DigitalSolutions.svg",
    FULL_STACK_DEVELOPMENT_SVG = "assets/illustrations/FullStackDevelopment.svg",
    CLOUD_INFRASTRUCTURE_SVG = "assets/illustrations/CouldInfrastructure.svg",
    UI_UX_DESIGN_SVG = "assets/illustrations/UIUXDesign.svg",
    PROJECT_ACQUIRE_CONVEYANCING = "assets/projects-images/acquire-conveyancing.png",
    PROJECT_ANGULAR_MASTER_PORTFOLIO = "assets/projects-images/angular-master-portfolio.png",
    PROJECT_ANGULAR_NEST_DOCKER = "assets/projects-images/angular-nest-docker.png",
    PROJECT_CUSTOM_DROPDOWN = "assets/projects-images/custom-dropdown.png",
    PROJECT_HMS_APP = "assets/projects-images/hms-app.png",
    PROJECT_KIWI_FINANCE = "assets/projects-images/kiwi-finance.png",
    PROJECT_LMS_APP = "assets/projects-images/lms-app.png",
    PROJECT_MUSIC_PLAYER = "assets/projects-images/music-player.png",
    PROJECT_RAS_FINANCE = "assets/projects-images/ras-finance.png",
    PROJECT_SA_TILING = "assets/projects-images/sa-tiling.png",
    PROJECT_THREEJS_IFC_VIEWER = "assets/projects-images/threejs-ifc-viewer.png",
    WORK_ACQUIRE_LOGO = "assets/experience-images/acquire-logo.png",
    WORK_ASITE_LOGO = "assets/experience-images/asite-logo.png",
    WORK_KIWI_LOGO = "assets/experience-images/kiwi-logo.png",
    WORK_SA_TILING_LOGO = "assets/experience-images/sa-tiling-logo.png",
    WORK_RAS_LOGO = "assets/experience-images/ras-logo.png",
    WORK_WORKXMATE_LOGO = "assets/experience-images/workxmate-logo.png",
    WORK_ADVENT_OF_CODE_LOGO = "assets/experience-images/advent-of-code-logo.jpeg",
    EDUCATION_GTU_LOGO = "assets/education-images/gtu-logo.png",
    EDUCATION_UNISA_LOGO = "assets/education-images/unisa-logo.png",
    PROFILE_PIC = "profile-images/profile-pic.png",
    ACHIEVEMENT_AOC_PIC = "assets/achievement-images/advent-of-code-finished.png",
    ACHIEVEMENT_CPC_RSP_WIN_PIC = "assets/achievement-images/cpc-rsp-win.jpg",
    ACHIEVEMENT_UNISA_CHANCELLORS_LETTER_2024_PIC = "assets/achievement-images/unisa-chancellors-letter-of-commandation-2024.png",
    SIMPLE_ICONS = "assets/icons",
}